+++ src/algorithms/mod.rs
@@ -1,0 +1,7 @@
+mod f;
+mod m;
+mod r;
+
+pub use f::F;
+pub use m::M;
+pub use r::R;