+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Unreleased
 ==========
 
  * **breaking:** Replaced `typenum` with const generics.
+ * `SamplingFrequency` internal field is no longer accessible.
 
 0.2.0
 ==========